Model-2/requirements.txt
numpy==1.25.2
pandas==2.1.0
scikit-learn>=1.3.3
nltk==3.9.3
spacy==3.6.1
gensim==4.3.1
faiss-cpu==1.7.4
umap-learn==0.5.11
polars==1.38.0
dask==2026.1.2
requests==2.32.1
orjson==3.9.9
matplotlib==3.8.1
seaborn==0.12.2
streamlit==1.30.0
Pillow==10.1.0
google-genai==0.2.0